Reduced the maximum number of spawned thread to not occupy all the resources of the GPU Changed the mini-batches size calculation

+++ sequential/sequential_version.cpp
@@ -3,7 +3,7 @@
 #include "k_means.h"
 #include "../utils/utils.cuh"
 
-void sequential_version(std::vector<std::vector<float>> data, int clusters, float max_tolerance, int max_iterations, std::mt19937 random_rng, int data_points_batch_size) {
+void sequential_version(std::vector<std::vector<double>> data, int clusters, double max_tolerance, int max_iterations, std::mt19937 random_rng, int data_points_batch_size) {
     if(data.size() < clusters) {
         std::cout << "There can't be more clusters than data points!" << std::endl;
         exit(1);
